Blooket Code to Hack
Copy this code (Ctrl+C), then open blooket. Right click and tap inspect, or Ctrl+Shift+I. Open console, which is located at the top of the screen on the right. Then do the command "Ctrl+V". If it says to type in "Allow pasting" or something related, do so. Now exit Inspect and scroll through cheats..





                     v
                     . .
                     0